Simulate permuted perturbation scores

Simulate null distributions of perturbation scores for each pathway through sample permutation.

Usage

generate_PermutedScore(logCPM, numOfTreat, NB = 1000, filePath, weight)

Arguments

logCPM

Matrix of normaslised logCPM where rows are genes and columns are samples. Row names need to be gene entrez IDs.

numOfTreat

Number of treatments (including control)

NB

Number of permutations

filePath

The file path to pathway topology matrices generated using function weightedAdjMatrix

weight

A vector of gene-wise weights derived from function weight_ssFC

Value

A list where each element is a vector of perturbation scores for a pathway.

Details

This function firstly randomly permute sample labels NB times to generate permuted logFCs, which are then used to compute permuted perturbation scores for each pathway. The function outputs a list that is of the same length as the list storing pathway topology matrices. Each element of the output list is for a pathway and contains a vector of permuted perturbation score of length NB. It's assumed that the permuted perturbation scores can be used to estimate the null distributions of perturbation scores.

The input logCPM matrix should be identical to the input used for weight_ssFC function. If the requested number of permutation (NB) is larger than the maximum number of permutations possible, NB will be set to the largest number of permutations possible instead.

Examples

require(AnnotationHub)
#> Loading required package: AnnotationHub
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: BiocFileCache
#> Loading required package: dbplyr
require(ensembldb)
#> Loading required package: ensembldb
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: GenomicFeatures
#> Loading required package: AnnotationDbi
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:AnnotationHub’:
#> 
#>     cache
#> Loading required package: AnnotationFilter
#> 
#> Attaching package: 'ensembldb'
#> The following object is masked from 'package:stats':
#> 
#>     filter
# convert rownamews of logCPM from gene ids to gene entrez IDs through `AnnotationHub`
ah <- AnnotationHub()
#> snapshotDate(): 2021-10-20
ah <- subset(ah,genome == "GRCh38" & title == "Ensembl 101 EnsDb for Homo sapiens")
ensDb <- ah[[1]]
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
rownames(logCPM_example) <- mapIds(ensDb, rownames(logCPM_example), "ENTREZID", keytype = "GENEID")
#> Warning: Unable to map 3137 of 18481 requested IDs.

# Remove genes that couldn't be matched to entrez IDs
logCPM_example <- logCPM_example[!is.na(rownames(logCPM_example)),]

#compute weighted single sample logFCs
ls <- weight_ssFC(logCPM_example, metadata = metadata_example,
 factor = "patient", control = "Vehicle")

# explore all species and databases supported by graphite
graphite::pathwayDatabases()
#>          species     database
#> 1      athaliana         kegg
#> 2      athaliana     pathbank
#> 3      athaliana wikipathways
#> 4        btaurus         kegg
#> 5        btaurus     pathbank
#> 6        btaurus     reactome
#> 7        btaurus wikipathways
#> 8       celegans         kegg
#> 9       celegans     pathbank
#> 10      celegans     reactome
#> 11      celegans wikipathways
#> 12   cfamiliaris         kegg
#> 13   cfamiliaris     reactome
#> 14   cfamiliaris wikipathways
#> 15 dmelanogaster         kegg
#> 16 dmelanogaster     pathbank
#> 17 dmelanogaster     reactome
#> 18 dmelanogaster wikipathways
#> 19        drerio         kegg
#> 20        drerio     reactome
#> 21        drerio wikipathways
#> 22         ecoli         kegg
#> 23         ecoli     pathbank
#> 24         ecoli wikipathways
#> 25       ggallus         kegg
#> 26       ggallus     reactome
#> 27       ggallus wikipathways
#> 28      hsapiens         kegg
#> 29      hsapiens      panther
#> 30      hsapiens     pathbank
#> 31      hsapiens     pharmgkb
#> 32      hsapiens     reactome
#> 33      hsapiens        smpdb
#> 34      hsapiens wikipathways
#> 35     mmusculus         kegg
#> 36     mmusculus     pathbank
#> 37     mmusculus     reactome
#> 38     mmusculus wikipathways
#> 39   rnorvegicus         kegg
#> 40   rnorvegicus     pathbank
#> 41   rnorvegicus     reactome
#> 42   rnorvegicus wikipathways
#> 43   scerevisiae         kegg
#> 44   scerevisiae     pathbank
#> 45   scerevisiae     reactome
#> 46   scerevisiae wikipathways
#> 47       sscrofa         kegg
#> 48       sscrofa     reactome
#> 49       sscrofa wikipathways
#> 50       xlaevis         kegg
weightedAdjMatrix(species = "hsapiens",
database = "kegg",
outputDir = "BminsI.rda")

permutedScore <- generate_PermutedScore(logCPM_example, numOfTreat = 2,
 NB = 100, filePath = "BminsI.rda", weight = ls$weight)