Normalise test perturbation scores by permutation

Normalise test perturbation scores by permutation

Usage

normaliseByPermutation(permutedScore, testScore, pAdj_method = "fdr")

Arguments

permutedScore

A list. Output of generate_PermutedScore

testScore

A matrix. Output of weight_ssFC

pAdj_method

Method for adjusting p-values for multiple comparisons. See ?p.adjust for methods available. Default to FDR.

Value

A dataframe.

Details

Normalise the test perturbation scores generated by weight_ssFC through the permuted perturbation scores derived from the generate_PermutedScore function. The mean absolute deviation(MAD) and median of perturbation scores for each pathway are firstly derived from the permuted perturbation scores. The test perturbation scores are then converted to robust z-scores using MADs and medians calculated.

Examples

if (FALSE) {
load(system.file("extdata", "gsTopology.rda", package = "SSPT"))
ssPertScore <- perturbationScore(ls$logFC, gsTopology)
permutedScore <- generate_PermutedScore(logCPM_example, numOfTreat = 2,
 NB = 100, gsTopology = gsTopology, weight = ls$weight)
normalisedScores <- normaliseByPermutation(permutedScore, ssPertScore)
 }
load(system.file("extdata", "normalisedScores.rda", package = "SSPT"))
head(normalisedScores)
#>                                     gs_name          MAD        MEDIAN sample
#> 1 EGFR tyrosine kinase inhibitor resistance 3.802026e-03  4.511294e-05     N1
#> 2                      Endocrine resistance 7.602171e-03 -3.622442e-04     N1
#> 3                     Antifolate resistance 3.314983e-04 -1.037652e-05     N1
#> 4                  Platinum drug resistance 2.673104e-03 -1.980562e-04     N1
#> 5                 mRNA surveillance pathway 3.490891e-05 -1.146214e-06     N1
#> 6                           RNA degradation 4.101007e-04  8.814987e-06     N1
#>              tA    robustZ    pvalue adjPvalue
#> 1 -3.838486e-03 -1.0214552 0.3070388 0.9978259
#> 2 -7.789896e-03 -0.9770436 0.3285476 0.9978259
#> 3 -9.488968e-05 -0.2549430 0.7987672 0.9978259
#> 4  7.921413e-04  0.3704298 0.7110623 0.9978259
#> 5 -8.951508e-06 -0.2235903 0.8230761 0.9978259
#> 6 -3.003849e-04 -0.7539609 0.4508727 0.9978259