Normalise test perturbation scores by permutation

Normalise test perturbation scores by permutation

Usage

normaliseByPermutation(permutedScore, testScore, pAdj_method = "fdr")

Arguments

permutedScore

A list. Output of generate_PermutedScore

testScore

A matrix. Output of weight_ssFC

pAdj_method

Method for adjusting p-values for multiple comparisons. See ?p.adjust for methods available. Default to FDR.

Value

A dataframe.

Details

Normalise the test perturbation scores generated by weight_ssFC through the permuted perturbation scores derived from the generate_PermutedScore function. The mean absolute deviation(MAD) and median of perturbation scores for each pathway are firstly derived from the permuted perturbation scores. The test perturbation scores are then converted to robust z-scores using MADs and medians calculated.

Examples

require(AnnotationHub)
require(ensembldb)
# convert rownamews of logCPM from gene ids to gene entrez IDs through `AnnotationHub`
ah <- AnnotationHub()
#> snapshotDate(): 2021-10-20
ah <- subset(ah,genome == "GRCh38" & title == "Ensembl 101 EnsDb for Homo sapiens")
ensDb <- ah[[1]]
#> loading from cache
rownames(logCPM_example) <- mapIds(ensDb, rownames(logCPM_example), "ENTREZID", keytype = "GENEID")
#> Warning: Unable to map 3137 of 18481 requested IDs.

# Remove genes that couldn't be matched to entrez IDs
logCPM_example <- logCPM_example[!is.na(rownames(logCPM_example)),]

#compute weighted single sample logFCs
ls <- weight_ssFC(logCPM_example, metadata = metadata_example,
 factor = "patient", control = "Vehicle")

# explore all species and databases supported by graphite
graphite::pathwayDatabases()
#>          species     database
#> 1      athaliana         kegg
#> 2      athaliana     pathbank
#> 3      athaliana wikipathways
#> 4        btaurus         kegg
#> 5        btaurus     pathbank
#> 6        btaurus     reactome
#> 7        btaurus wikipathways
#> 8       celegans         kegg
#> 9       celegans     pathbank
#> 10      celegans     reactome
#> 11      celegans wikipathways
#> 12   cfamiliaris         kegg
#> 13   cfamiliaris     reactome
#> 14   cfamiliaris wikipathways
#> 15 dmelanogaster         kegg
#> 16 dmelanogaster     pathbank
#> 17 dmelanogaster     reactome
#> 18 dmelanogaster wikipathways
#> 19        drerio         kegg
#> 20        drerio     reactome
#> 21        drerio wikipathways
#> 22         ecoli         kegg
#> 23         ecoli     pathbank
#> 24         ecoli wikipathways
#> 25       ggallus         kegg
#> 26       ggallus     reactome
#> 27       ggallus wikipathways
#> 28      hsapiens         kegg
#> 29      hsapiens      panther
#> 30      hsapiens     pathbank
#> 31      hsapiens     pharmgkb
#> 32      hsapiens     reactome
#> 33      hsapiens        smpdb
#> 34      hsapiens wikipathways
#> 35     mmusculus         kegg
#> 36     mmusculus     pathbank
#> 37     mmusculus     reactome
#> 38     mmusculus wikipathways
#> 39   rnorvegicus         kegg
#> 40   rnorvegicus     pathbank
#> 41   rnorvegicus     reactome
#> 42   rnorvegicus wikipathways
#> 43   scerevisiae         kegg
#> 44   scerevisiae     pathbank
#> 45   scerevisiae     reactome
#> 46   scerevisiae wikipathways
#> 47       sscrofa         kegg
#> 48       sscrofa     reactome
#> 49       sscrofa wikipathways
#> 50       xlaevis         kegg
weightedAdjMatrix(species = "hsapiens",
database = "kegg",
outputDir = "BminsI.rda")

# compute test perturbation scores
ssPertScore <- perturbationScore(ls$logFC, "BminsI.rda")
# generate permuted perturbation scores
permutedScore <- generate_PermutedScore(logCPM_example, numOfTreat = 2,
 NB = 100, filePath = "BminsI.rda", weight = ls$weight)
normalisedScores <- normaliseByPermutation(permutedScore, ssPertScore)